Define corpus
The corpus is a collection of valid and invalid files that may be used for
development or, as in the knowledge base, as a data source for
verification and reproducibility.

Data model of the corpus is the same as in the knowledge base but
considers only data sources; file and format objects. Most of preliminary error
objects with full validator output from the invalid files are included. The
precise error messages for the knowledge base are not defined in the
preliminary error objects.

The knowledge base is located in database/ and the corpus in
corpus/. The data are in a <objects>.jsonl files that may or may not be indexed
to directories.

The corpus is collected with file-scraper 2.4.0 from file-scraper 2.4.0
tests/data/.