Prepare errors from corpus/errors.jsonl for migration
The errors to be migrated to the knowledge base are stored to 'pre' files in
the validator index directories (database/solutions/<validator>/pre.

Relevant error message section from the validator output example should
be picked for the error that is being prepared. Then, if the same error
message is raised also from other files, the picked (defined) error message
should refer to those files as well from the .files[] field.