Migrate MPO files to JPEG with ExifTool
General file format description for JPEG is added to formats.jsonl.
Descriptions from 'virtual' files are removed; the wellFormed field
is enough plus the generic 'virtual' files should be removed or at
least replaced with objetcs that have a location.

Temporary 'list' files were used in migration of a corpus to the knowledge
base and are removed.